refactor: consolidate screenshot utilities to single location
Remove the old monolithic screenshot-utils.ts and standardize on the
adapter pattern for all screenshot utility usage. The interface now
lives in the interfaces directory, and ViewHierarchy uses the adapter
singleton instead of direct static class calls.

Closes #1121
