+++ src/utils/interfaces/index.ts
@@ -14,7 +14,8 @@ export { ScreenshotComparator } from "../screenshot/ScreenshotComparator";
 export { PerceptualHasher } from "../screenshot/PerceptualHasher";
 export { ScreenshotCache } from "../screenshot/ScreenshotCache";
 export { ScreenshotMatcher } from "../screenshot/ScreenshotMatcher";
-export { ScreenshotUtils } from "../screenshot/ScreenshotUtils";
+export { ScreenshotUtils as ScreenshotUtilsImpl } from "../screenshot/ScreenshotUtils";
+export type { ScreenshotUtils } from "./ScreenshotUtils";
 export { screenshotUtilsAdapter } from "../ScreenshotUtilsAdapter";
 export { ImageUtils } from "./ImageUtils";
 export { PlanUtils } from "./PlanUtils";
